Try to fix mastermind code (explanation of problem in the long message) The fix is to make sure only one set of (tumblers, gears) is applied for each play of the instrument.
The problem is invasive. There's a microscopic delay (unnoticeable most of the
time, unless under high lag) between one message and the next being printed
(this includes --More-- prompts, since NH has to find out that it needs to
issue a --More--). So what's happening is:

* We receive "You hear 3 gears turn."
* The code updates the internal VT so the top line reads the message we got
* The code runs the mastermind code, which sees the "You hear 3 gears turn."
  and updates its internal state
* Small pause as NetHack figures out what happens for the rest of the turn
* If nothing else generates output, the code worked fine
* If you got a --More-- prompt (or indeed any other output) the code would
  see the new text, update the internal VT, and rerun the code
* The mastermind code again sees "You hear 3 gears turn." but with a --More--
  afterwards (which it discards); updates its internal state again, which
  effectively means it thinks there were two "3 gear" responses

A general fix is pretty difficult, because sometimes the --More-- might be in
the same packet (and so the appropriate code would correctly only be run once),
or, indeed, the --More-- might not show up at all.

One fix though is making code sensitive to these kinds of things (such as the
mysterious force counter) run only against the input stream, not against the VT.

Any ideas?